+++ src/main/kotlin/icu/windea/pls/ep/config/CwtOverriddenConfigProviders.kt
@@ -21,7 +21,8 @@ class CwtSwitchOverriddenConfigProvider : CwtOverriddenConfigProvider {
     
     @Suppress("UNCHECKED_CAST")
     override fun <T : CwtMemberConfig<*>> getOverriddenConfigs(contextElement: PsiElement, config: T): List<T>? {
-        //重载switch = {...}和inverted_switch = {...}中对应的CWT规则为scalar的属性的键对应的CWT规则
+        //重载switch = {...}中对应的CWT规则为scalar的属性的键对应的CWT规则
+        //重载inverted_switch = {...}中对应的CWT规则为scalar的属性的键对应的CWT规则
         //兼容使用内联或者使用封装变量的情况
         if(config !is CwtPropertyConfig) return null
         if(config.key != Data.CASE_KEY) return null